animation in HTML+JS using pngs as frames
Doesn't work great: choppy and a bit RAM-heavy in Firefox,
very RAM-heavy in Opera for a thousand frames or so.

Might be better to try to use <video> tag instead of rewriting
an animation engine badly. If that won't look nice enough,
try to tweak the Javascript to work better.