improve process.py performance a bit
It now takes about 14 seconds to load in an hour (1440 files) of car2go
data on my machine.

json.load() is a big issue at this point, taking about half of the
processing time. Not really sure how this can be improved short of rewriting
everything.

Could for instance only read in all the data files separately on the
first run, then process the data into lists of parking positions and trips
and save that - should be a lot smaller. But that's a big change.

simplejson is the best performing of json, simplejson, ujson in loading
a large list of dicts in my tests.