Improve process.py process_data loop
Rewriting process_data to more explicitly keep track of finished and
unfinished rips and parking periods.

This will allow more efficient storage and processing since we can store
just trips and parking periods rather than every parked car's position
as we currently get out of the APIs.

The new code has the same output as previous code EXCEPT fuel use
is slightly different, due to a bug.

The old code actually has a bug that causes it to discard valid
round-trips. I've noted this in a comment - I am not correcting it now
so that I can verify that the rewrite behaves the same as the code it
replaces.

The fuel problem affects 0.44% of all trips. Considering that fuel use
data is not particularly reliable in the first place (e.g. gas car2go
resolution is about 3%), this isn't a big issue.

For now the changes are limited to process_data and batch_load_data.
With the new system in place, I'll be able to rewrite the rest to use
the new data format in a subsequent change.

As an added bonus this code is also ~2x faster, old code took about
14-17 s to read in 1440 data files and new code takes about 8 s.