don't copy when skipping files in batch_load_data
Don't copy the missing files, load them in as and when needed.

I think this was necessary at one point but with changes on how files
are loaded it long since hasn't been.

Prevents us from thinking we have all the files on subsequent runs -
missing files will stay missing.

Technically this will also be more correct as we're not pretending that
our data is from a file that we don't have.