separate timers in website and downloader
They used to use cars.py's common one. That was not very good. This is
slightly less bad.

There is basically no difference between timing a download in
download.get_url() and in get_all_cars_text() - any internal Python time
to traverse two functions is dwarfed by the HTTP request. The only thing
we're possibly losing is the breakdown of HEAD and GET requests where
necessary, but currently this is only for Enjoy which doesn't even have
electric cars, so it doesn't matter.