simplify systems files
The current structure dates down to when car2go was the only system
supported and its information was in city.py in the repo root.

We no longer need the CITIES info to be in a separate file, and the
division between __init__.py and city.py was largely artificial for most
systems. I've kept it for car2go and Drivenow as their city.py is a bit
more complex.

Also simplify a whole bunch of things where I was trying to be too smart
with using lambdas or looping over a dictionary containing one item.