remove homemade timer/profiler from process.py
This is equivalent to what was done for normalize.py in 3411a2f7a8.

The "debug" option printed a bunch of timing information. This is easily
recreated and improved upon by using cProfile. Removing a global
"timer" variable and print() calls will help with parallelization.

Plus, the TL;DR of both old timer and cProfile is that
matplotlib.figure.savefig() takes about 75% of the time, so that's the
first place to look for speedups - though I understand this includes
actually rendering the figure.

Also update comments that mentioned how long something took a long time
ago - they were mostly irrelevant or outdated by now.