Remove test set limit in GenerateTest
With the speed-up in c865a7d0 tests now run fast enough that this isn't
necessary. Most of the time is still spent actually opening up the tgz
and in particular dumping the JSON, so returning to originally intended
comparison window is not a problem.

On my mediocre computer this change added max 20 seconds to the test
(23 to 42 seconds, but usually under 30 seconds now).