fix tests in Python 2; includes bonus fix of years-old bug
All tests now pass under Python 2.7.8.

Force float division with `fuel / 100.0`, we get test failures in
Python 2 otherwise.

Python 2 has no tempfile.TemporaryDirectory(), the quickest way to deal
with that was to make my own using tempfile.mkdtemp().

While running tests in Python 2 I found an interesting bug, one of these
things where you wonder how it ever worked in a first place. Normalizing
never worked if the path to the source file had any uppercase characters
in it. This was due to a call to lower(), which was first introduced to
process.py in 8433b5244d65 in January 2013 and somehow was kept all
throughout the refactors and moves, ultimately ending up in normalize.py
in 50a5d2d6d32a. It serves no purpose now as far as I can tell -
filenames are kept in sync with city names in download.py.
I think it was originally it was held over from back when we got
city name in HTTP URL param.

This didn't come up until now because all of my paths are lowercase,
and because as an implementation detail on Linux, Python 3 temporary
directories are created with all-lowercase names. Python 2 appears to
create temporary directories with a mix of uppercase and lowercase,
which I was needlessly lowercasing.